--- TeamCode/src/main/java/FieldCentricMecanumTeleOpProgresive.java
@@ -1,154 +1,0 @@
-import static org.firstinspires.ftc.teamcode.SubMotor.target;
-
-import com.acmerobotics.dashboard.FtcDashboard;
-import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
-import com.arcrobotics.ftclib.controller.PIDController;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.Gamepad;
-import com.qualcomm.robotcore.hardware.IMU;
-import com.qualcomm.robotcore.hardware.Servo;
-
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.CurrentUnit;
-import org.firstinspires.ftc.teamcode.SubMotor;
-
-
-@TeleOp
-public class FieldCentricMecanumTeleOpProgresive extends LinearOpMode {
-
-    public DcMotorEx slidemotorright;
-    public DcMotorEx slidemotorleft;
-    
-
-
-    private Servo SlideR;
-    private Servo SliderL;
-    private Servo AlongeR;
-    private Servo AllongeL;
-    private Servo pince;
-    private Servo bucket;
-    private Servo pinceArriere;
-    private Servo poignet;
-    private SubMotor subMotor =null;
-
-    public void runOpMode(){
-
-        Gamepad currentGamepad1 = new Gamepad();
-        Gamepad currentGamepad2 = new Gamepad();
-
-        Gamepad previousGamepad1 = new Gamepad();
-        Gamepad previousGamepad2 = new Gamepad();
-
-
-        boolean intakeToggle = false;
-        //boolean poignettoggle = true;
-        boolean pinceToggle = false;
-
-
-        slidemotorright = hardwareMap.get(DcMotorEx.class, "slidemotorright");
-        slidemotorleft = hardwareMap.get(DcMotorEx.class, "slidemotorleft");
-
-
-        // Declare our motors
-        // Make sure your ID's match your configuration
-        DcMotorEx leftFront = hardwareMap.get(DcMotorEx.class,"leftFront");
-        DcMotor rightBack = hardwareMap.dcMotor.get("rightBack");
-        DcMotor rightFront = hardwareMap.dcMotor.get("rightFront");
-        DcMotor leftBack = hardwareMap.dcMotor.get("leftBack");
-
-
-        leftBack.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        leftFront.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        rightFront.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        rightBack.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-
-        //servo
-        SlideR = hardwareMap.get(Servo.class, "SlideR");
-        SliderL = hardwareMap.get(Servo.class, "SliderL");
-        AlongeR = hardwareMap.get(Servo.class, "AlongeR");
-        pince = hardwareMap.get(Servo.class, "pince");
-        bucket = hardwareMap.get(Servo.class, "bucket");
-        AllongeL = hardwareMap.get(Servo.class, "AllongeL");
-        pinceArriere = hardwareMap.get(Servo.class, "pinceArriere");
-        poignet = hardwareMap.get(Servo.class, "poignet");
-        SlideR.setDirection(Servo.Direction.REVERSE);
-        AllongeL.setDirection(Servo.Direction.REVERSE);
-
-        leftBack.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        leftFront.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        rightBack.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        rightFront.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        // Reverse the right side motors. This may be wrong for your setup.
-        // If your robot moves backwards when commanded to go forwards,
-        // reverse the left side instead.
-        // See the note about this earlier on this page.
-        leftBack.setDirection(DcMotorSimple.Direction.REVERSE);
-        rightBack.setDirection(DcMotorSimple.Direction.REVERSE);
-
-        // Retrieve the IMU from the hardware map
-        IMU imu = hardwareMap.get(IMU.class, "imu");
-        // Adjust the orientation parameters to match your robot
-        IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
-                RevHubOrientationOnRobot.LogoFacingDirection.LEFT,
-                RevHubOrientationOnRobot.UsbFacingDirection.UP));
-        imu.resetYaw();
-        // Without this, the REV Hub's orientation is assumed to be logo up / USB forward
-        imu.initialize(parameters);
-
-
-        waitForStart();
-
-        while (opModeIsActive()) {
-
-            double y = -gamepad1.left_stick_y; // Remember, Y stick value is reversed
-            double x = gamepad1.left_stick_x;
-            double rx = gamepad1.right_stick_x;
-
-            // This button choice was made so that it is hard to hit on accident,
-            // it can be freely changed based on preference.
-            // The equivalent button is start on Xbox-style controllers.
-            if (gamepad1.options) {
-                imu.resetYaw();
-            }
-
-            double botHeading = imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.RADIANS);
-
-            // Rotate the movement direction counter to the bot's rotation
-            double rotX = x * Math.cos(-botHeading) - y * Math.sin(-botHeading);
-            double rotY = x * Math.sin(-botHeading) + y * Math.cos(-botHeading);
-
-            rotX = rotX * 1.1;  // Counteract imperfect strafing
-
-            // Denominator is the largest motor power (absolute value) or 1
-            // This ensures all the powers maintain the same ratio,
-            // but only if at least one is out of the range [-1, 1]
-            double denominator = Math.max(Math.abs(rotY) + Math.abs(rotX) + Math.abs(rx), 1);
-            double frontLeftPower = (rotY + rotX + rx) / denominator;
-            double backLeftPower = (rotY - rotX + rx) / denominator;
-            double frontRightPower = (rotY - rotX - rx) / denominator;
-            double backRightPower = (rotY + rotX - rx) / denominator;
-
-            leftFront.setPower(frontLeftPower);
-            leftBack.setPower(backLeftPower);
-            rightFront.setPower(frontRightPower);
-            rightBack.setPower(backRightPower);
-            //ceci est un test;
-
-
-            previousGamepad1.copy(currentGamepad1);
-            currentGamepad1.copy(gamepad1);
-
-
-
-        }
-    }
-}
-
-
